+++ packages/compat/uni-builder/src/webpack/index.ts
@@ -4,6 +4,7 @@ import {
   type RsbuildPlugin,
   type RsbuildInstance,
 } from '@rsbuild/core';
+import type { RsbuildTarget } from '@rsbuild/shared';
 import type { UniBuilderWebpackConfig } from '../types';
 import type { CreateWebpackBuilderOptions } from '../types';
 import { parseCommonConfig } from '../shared/parseCommonConfig';
@@ -17,6 +18,7 @@ export async function parseConfig(
   uniBuilderConfig: UniBuilderWebpackConfig,
   cwd: string,
   frameworkConfigPath?: string,
+  target?: RsbuildTarget | RsbuildTarget[],
 ): Promise<{
   rsbuildConfig: RsbuildConfig;
   rsbuildPlugins: RsbuildPlugin[];
@@ -25,6 +27,7 @@ export async function parseConfig(
     uniBuilderConfig,
     cwd,
     frameworkConfigPath,
+    target,
   );
 
   rsbuildPlugins.push(pluginBabel(uniBuilderConfig.tools?.babel));
@@ -74,11 +77,14 @@ export async function createWebpackBuilder(
     withDefaultConfig(options.config),
     options.cwd,
     options.frameworkConfigPath,
+    options.target,
   );
+
   const { webpackProvider } = await import('@rsbuild/webpack');
   const rsbuild = await createRsbuild({
     rsbuildConfig,
     provider: webpackProvider,
+    target: options.target || 'web',
   });
 
   rsbuild.addPlugins([