+++ packages/core/src/plugins/index.ts
@@ -1,6 +1,7 @@
 import type { Plugins } from '@rsbuild/shared';
 
 export const plugins: Plugins = {
+  basic: () => import('./basic').then((m) => m.pluginBasic()),
   html: () => import('./html').then((m) => m.pluginHtml()),
   cleanOutput: () => import('./cleanOutput').then((m) => m.pluginCleanOutput()),
   startUrl: () => import('./startUrl').then((m) => m.pluginStartUrl()),
@@ -26,5 +27,4 @@ export const plugins: Plugins = {
     import('./preloadOrPrefetch').then((m) => m.pluginPreloadOrPrefetch()),
   performance: () => import('./performance').then((m) => m.pluginPerformance()),
   define: () => import('./define').then((m) => m.pluginDefine()),
-  server: () => import('./server').then((m) => m.pluginServer()),
 };