+++ packages/babel-preset/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rsbuild/babel-preset
 
+## 0.1.9
+
+### Patch Changes
+
+- @rsbuild/plugin-babel@0.1.9
+- @rsbuild/shared@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes